Move CustomModal to Modals folder
