+++ src/js/Core/Modals/CustomModal.ts
